refactored AC namespaces
